pyproject.toml
[tool.poetry]
name = "repl_python3_RPICDB-Dashboard"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Quart = "^0.15.1"
pymongo = {version = "^3.12.0", extras = ["srv"]}
"discord.py" = "^1.7.3"
Quart-Discord = "^2.1.4"
python-youtube = "^0.8.1"

[tool.poetry.extras]
srv = ["pymongo"]

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
